# Conflicts:
#	.idea/dataSources/da3d623e-4e46-4f59-a223-ea8186a79bf3.xml

+++ src/main/java/com/xq/crowd_funding/raisefunding/beans/pojo/TMemberLaunchInfoPO.java
@@ -5,8 +5,9 @@ import lombok.Data;
 @Data
 public class TMemberLaunchInfoPO {
 
-  private Long id;
-  private Long memberid;
+  private Integer id;
+  private Integer memberid;
+  // 简单介绍
   private String descriptionSimple;
   private String descriptionDetail;
   private String phoneNum;
